Rewrite unit tests under fee, hashing directory
